change: switch to xxh3_128 hashes

+++ src/job.rs
@@ -29,8 +29,8 @@ pub enum JobType {
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Proof {
-    // xxh3_64
-    pub hash: u64,
+    // xxh3
+    pub hash: u128,
 
     pub blob: Vec<u8>
 }
